Move utility functions
